Formulario e tela de barrada de vida
Formulario e tela de barrada de vida

+++ Assets/Exemplo 2/Script/Creature.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class Creature : MonoBehaviour {
 
-	public int vida;
+	public float vida;
 	public int forcaFisica;
 	public int defesaFisica;
 	public int forcaMagica;
